backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
networkx==3.2.1
numpy==1.24.3
transformers>=4.45.0
accelerate>=0.26.0
bitsandbytes>=0.43.0
huggingface-hub
sentencepiece
sqlalchemy
aiosqlite
python-dotenv
gitpython